// WADNR.Web/src/app/shared/generated/model/diff-section.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DiffField } from './diff-field';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class DiffSection { 
    Title?: string | null;
    Type?: string | null;
    Fields?: Array<DiffField> | null;
    Headers?: Array<string> | null;
    OriginalRows?: Array<Array<string>> | null;
    UpdatedRows?: Array<Array<string>> | null;
    OriginalItems?: Array<string> | null;
    UpdatedItems?: Array<string> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface DiffSectionForm { 
    Title?: FormControl<string>;
    Type?: FormControl<string>;
    Fields?: FormControl<Array<DiffField>>;
    Headers?: FormControl<Array<string>>;
    OriginalRows?: FormControl<Array<Array<string>>>;
    UpdatedRows?: FormControl<Array<Array<string>>>;
    OriginalItems?: FormControl<Array<string>>;
    UpdatedItems?: FormControl<Array<string>>;
}

export class DiffSectionFormControls { 
    public static Title = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static Type = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static Fields = (value: FormControlState<Array<DiffField>> | Array<DiffField> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<DiffField>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static Headers = (value: FormControlState<Array<string>> | Array<string> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<string>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static OriginalRows = (value: FormControlState<Array<Array<string>>> | Array<Array<string>> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<Array<string>>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static UpdatedRows = (value: FormControlState<Array<Array<string>>> | Array<Array<string>> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<Array<string>>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static OriginalItems = (value: FormControlState<Array<string>> | Array<string> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<string>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static UpdatedItems = (value: FormControlState<Array<string>> | Array<string> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<string>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}